Renamed event folder to entities folder
