+++ phase0/specification.md
@@ -2,9 +2,9 @@
 
 [specification on google docs that's underline](https://docs.google.com/document/d/1n7Uq_OP2d5JIQmQLOnWaaJ42P9wVLEaM4yTghBf5wIA/edit?usp=sharing)
 
-A calendar in which a user can **create, edit, or delete** **birthday events** that correspond to certain dates and certain friends.
+A calendar in which a user can **create, edit, or delete** **birthday entities** that correspond to certain dates and certain friends.
 Users can set a time before the **event** **to be notified**. Each **birthday event** should be able to **include personal comments and gift ideas**.
-Users should have the option to **look back** at previous events or be **reminded** of them in subsequent years.
+Users should have the option to **look back** at previous entities or be **reminded** of them in subsequent years.
 
 ## Example usage
