+++ trackr/src/main/java/entities/Birthday.java
@@ -1,4 +1,4 @@
-package events;
+package entities;
 
 import java.util.Date;
