+++ trackr/src/test/java/entities/AnniversaryTest.java
@@ -1,0 +1,5 @@
+package entities;
+
+class AnniversaryTest {
+
+}
