--- trackr/src/test/java/events/AnniversaryTest.java
@@ -1,7 +1,0 @@
-package events;
-
-import static org.junit.jupiter.api.Assertions.*;
-
-class AnniversaryTest {
-
-}
