--- trackr/src/test/java/events/BirthdayTest.java
@@ -1,7 +1,0 @@
-package events;
-
-import static org.junit.jupiter.api.Assertions.*;
-
-class BirthdayTest {
-
-}
